requirements.txt
streamlit
pickle
pillow
sklearn
